requirements.txt
cassandra-driver
pandas
matplotlib
tqdm